+++ CODEOWNERS
@@ -1,3 +1,3 @@
 # These owner(s) will be the default owners for everything in the repo.
 
-* @agitrubard @yvzcnr @Nbhtozcn @cagla88
+* @agitrubard @yvzcnr @Nbhtozcn @cagla88 @TekinMerve
