+++ src/test/java/org/ays/feature/UsersDelete.feature
@@ -1,4 +1,4 @@
-@Ignored
+@Ignore
 Feature: Testing Users
 
   Background:
