+++ src/test/java/org/ays/feature/UsersShow.feature
@@ -1,4 +1,4 @@
-@Ignored
+@Regression
 Feature: Testing Users
 
   Background:
